+++ changelog.md
@@ -1,1 +1,2 @@
 - Added platform fluid tags.
+- Fixed Ad Astra using a shared property object for all metal blocks, causing them to have properties like no occlusion when they should have it.